chore: update minimum Elixir version, Styler, and action workflows (#95)
* chore: upgrade Styler

styler 1.0.0 dropped support for Elixir 1.14 and there are code changes
present because of additional style updates provided. As such, minimum
Elixir version is being bumped to Elixir 1.15.

* chore: Improve lint-commit action security

Running `actionlint` reports:

```
.github/workflows/lint-commit.yaml:20:24:
"github.event.pull_request.title" is potentially untrusted. avoid using
it directly in inline scripts. instead, pass it through an environment
variable. see
https://docs.github.com/en/actions/security-guides/security-hardening-for-github-actions
for more details [expression]
```

* chore: Update ci workflow

The tested Elixirs are 1.15–1.17 on OTP 24–27 (as appropriate), all
running on ubuntu-latest.

- Updated actions/checkout to v4
- Updated actions/cache to v4, modified the keys to use the
  erlef/setup-beam outputs for the matched elixir and otp versions
  rather than the Matrix specification. This will caus a little more
  cache churn when new versions are released, but will prevent possible
  cache *errors* if incompatibilities are present.

Modified the `formatter` job to use a matrix so that various display and
cache key markers are updated when the format runner is updated.

* chore: update release workflow

Mostly already updated, but added the following changes:

- Updated actions/checkout to v4
- Updated actions/cache to v4, modified the keys to use the
  erlef/setup-beam outputs for the matched elixir and otp versions
  rather than the Matrix specification. This will caus a little more
  cache churn when new versions are released, but will prevent possible
  cache *errors* if incompatibilities are present.

* chore: Update .rtx.toml

This should *probably* be renamed to `.mise.toml` since the tool name
changed, but the old name still works for people who haven't yet changed
from `rtx` to `mise`.