fix: allow atom layout module names
String values for layout module names are required to be supported in
the front matter, but it feels odd to require that they be strings in
config. This enables the specification of layout modules as atoms and
fixes a bug in PageExtension which required a layout value in the
front matter at all times.